requirements.txt
aiogram==3.4.1
psycopg2-binary==2.9.9
requests==2.31.0